big chonky refactor
