--- Backend/auth/tests/main.go
@@ -1,3 +1,0 @@
-package tests
-
-// TODO :P
