+++ Backend/database/repositories/main.go
@@ -8,3 +8,20 @@ var context contexts.DatabaseContext
 func init() {
 	context = contexts.GetDatabaseContext()
 }
+
+// enum of repositories
+const (
+	FILESYSTEM = iota
+)
+
+// small factory for setting up and returning a repository
+func GetRepository(repo int) interface{} {
+	switch repo {
+	case FILESYSTEM:
+		return FilesystemRepository{
+			context,
+		}
+	default:
+		return nil
+	}
+}
