+++ Backend/endpoints/filesystem_endpoints.go
@@ -1,4 +1,4 @@
-package filesystem
+package endpoints
 
 import (
 	"encoding/json"
